refactor(backup/gcs): use custom exception type for casting errors
