+++ backup-stores/gcs/src/main/java/io/camunda/zeebe/backup/gcs/GcsBackupStoreException.java
@@ -7,6 +7,8 @@
  */
 package io.camunda.zeebe.backup.gcs;
 
+import io.camunda.zeebe.backup.gcs.manifest.StatusCode;
+
 public abstract class GcsBackupStoreException extends RuntimeException {
   public GcsBackupStoreException(final String message) {
     super(message);
@@ -19,6 +21,12 @@ public abstract class GcsBackupStoreException extends RuntimeException {
   public static class InvalidPersistedManifestState extends GcsBackupStoreException {
     public InvalidPersistedManifestState(final String errorMessage) {
       super(errorMessage);
+    }
+  }
+
+  public static class UnexpectedManifestState extends GcsBackupStoreException {
+    public UnexpectedManifestState(final StatusCode expected, final StatusCode actual) {
+      super("Expected manifest in state '%s', but was in '%s'".formatted(expected, actual));
     }
   }
