+++ backup-stores/gcs/src/main/java/io/camunda/zeebe/backup/gcs/manifest/ManifestImpl.java
@@ -7,39 +7,37 @@
  */
 package io.camunda.zeebe.backup.gcs.manifest;
 
-import static io.camunda.zeebe.backup.gcs.manifest.BackupStatusCode.COMPLETED;
-import static io.camunda.zeebe.backup.gcs.manifest.BackupStatusCode.FAILED;
-import static io.camunda.zeebe.backup.gcs.manifest.BackupStatusCode.IN_PROGRESS;
+import static io.camunda.zeebe.backup.gcs.manifest.StatusCode.COMPLETED;
+import static io.camunda.zeebe.backup.gcs.manifest.StatusCode.FAILED;
+import static io.camunda.zeebe.backup.gcs.manifest.StatusCode.IN_PROGRESS;
 
 import io.camunda.zeebe.backup.common.BackupDescriptorImpl;
 import io.camunda.zeebe.backup.common.BackupIdentifierImpl;
 import io.camunda.zeebe.backup.gcs.GcsBackupStoreException.InvalidPersistedManifestState;
+import io.camunda.zeebe.backup.gcs.GcsBackupStoreException.UnexpectedManifestState;
 import java.time.Instant;
 
 record ManifestImpl(
     BackupIdentifierImpl id,
     BackupDescriptorImpl descriptor,
-    BackupStatusCode statusCode,
+    StatusCode statusCode,
     Instant createdAt,
     Instant modifiedAt,
     String failureReason)
     implements InProgressManifest, CompletedManifest, FailedManifest {
 
-  public static final String ERROR_MSG_AS_CAST = "Expected to be in '%s' state, but was in '%s'";
-  public static final String ERROR_MSG_CREATION_FAILURE =
-      "Expected to set failureReason '%s', with status code 'FAILED' but was '%s'";
-
   ManifestImpl {
     if (failureReason != null && statusCode != FAILED) {
-      final var errorMessage = String.format(ERROR_MSG_CREATION_FAILURE, failureReason, statusCode);
-      throw new InvalidPersistedManifestState(errorMessage);
+      throw new InvalidPersistedManifestState(
+          "Manifest in state '%s' must be 'FAILED to have have failureReason '%s'"
+              .formatted(statusCode, failureReason));
     }
   }
 
   ManifestImpl(
       final BackupIdentifierImpl id,
       final BackupDescriptorImpl descriptor,
-      final BackupStatusCode statusCode,
+      final StatusCode statusCode,
       final Instant createdAt,
       final Instant modifiedAt) {
     this(id, descriptor, statusCode, createdAt, modifiedAt, null);
@@ -58,8 +56,7 @@ record ManifestImpl(
   @Override
   public InProgressManifest asInProgress() {
     if (statusCode != IN_PROGRESS) {
-      final String errorMsg = String.format(ERROR_MSG_AS_CAST, IN_PROGRESS, statusCode);
-      throw new IllegalStateException(errorMsg);
+      throw new UnexpectedManifestState(IN_PROGRESS, statusCode);
     }
 
     return this;
@@ -68,8 +65,7 @@ record ManifestImpl(
   @Override
   public CompletedManifest asCompleted() {
     if (statusCode != COMPLETED) {
-      final String errorMsg = String.format(ERROR_MSG_AS_CAST, COMPLETED, statusCode);
-      throw new IllegalStateException(errorMsg);
+      throw new UnexpectedManifestState(COMPLETED, statusCode);
     }
 
     return this;
@@ -78,8 +74,7 @@ record ManifestImpl(
   @Override
   public FailedManifest asFailed() {
     if (statusCode != FAILED) {
-      final String errorMsg = String.format(ERROR_MSG_AS_CAST, FAILED, statusCode);
-      throw new IllegalStateException(errorMsg);
+      throw new UnexpectedManifestState(FAILED, statusCode);
     }
 
     return this;
